Update based on Diego's review
